refact: alterar nome do pacote
