+++ fase-03/src/main/java/br/com/fiap/techchallenge/infra/exceptions/errors/ControllerNotFoundException.java
@@ -1,0 +1,8 @@
+package br.com.fiap.techchallenge.infra.exceptions.errors;
+
+public class ControllerNotFoundException extends RuntimeException {
+
+    public ControllerNotFoundException(String message){
+        super(message);
+    }
+}
