+++ fase-03/src/main/java/br/com/fiap/techchallenge/utils/StringSanitizer.java
@@ -1,0 +1,8 @@
+package br.com.fiap.techchallenge.utils;
+
+public class StringSanitizer {
+
+    public static String somenteNumeros(String string) {
+        return string.replaceAll("[^0-9]", "");
+    }
+}
