--- fase-03/src/main/java/com/example/techChallengeParkimetro/TechChallengeParkimetroApplication.java
@@ -1,15 +1,0 @@
-package com.example.techChallengeParkimetro;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.scheduling.annotation.EnableScheduling;
-
-@SpringBootApplication
-@EnableScheduling
-public class TechChallengeParkimetroApplication {
-
-    public static void main(String[] args) {
-        SpringApplication.run(TechChallengeParkimetroApplication.class, args);
-    }
-
-}
