--- fase-03/src/main/java/com/example/techChallengeParkimetro/entities/dtos/VeiculoDTO.java
@@ -1,18 +1,0 @@
-package com.example.techChallengeParkimetro.entities.dtos;
-
-import com.example.techChallengeParkimetro.entities.Veiculo;
-import jakarta.validation.constraints.NotBlank;
-
-public record VeiculoDTO(
-        @NotBlank
-        String marca,
-        @NotBlank
-        String modelo,
-        @NotBlank
-        String placa,
-        String condutorCpf
-) {
-    public Veiculo toEntity(String condutorCpf) {
-        return new Veiculo(marca, modelo, placa.toUpperCase(), condutorCpf);
-    }
-}
