--- fase-03/src/main/java/com/example/techChallengeParkimetro/infra/exceptions/errors/ControllerNotFoundException.java
@@ -1,8 +1,0 @@
-package com.example.techChallengeParkimetro.infra.exceptions.errors;
-
-public class ControllerNotFoundException extends RuntimeException {
-
-    public ControllerNotFoundException(String message){
-        super(message);
-    }
-}
