--- fase-03/src/main/java/com/example/techChallengeParkimetro/utils/StringSanitizer.java
@@ -1,8 +1,0 @@
-package com.example.techChallengeParkimetro.utils;
-
-public class StringSanitizer {
-
-    public static String somenteNumeros(String string) {
-        return string.replaceAll("[^0-9]", "");
-    }
-}
