+++ fase-03/src/test/java/br/com/fiap/techchallenge/ParkimetroApplicationTests.java
@@ -1,0 +1,13 @@
+package br.com.fiap.techchallenge;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class ParkimetroApplicationTests {
+
+	@Test
+	void contextLoads() {
+	}
+
+}
