--- fase-03/src/test/java/com/example/techChallengeParkimetro/TechChallengeParkimetroApplicationTests.java
@@ -1,13 +1,0 @@
-package com.example.techChallengeParkimetro;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class TechChallengeParkimetroApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
